delete ethereum example move the TaintTrackingWithHistory algorithm into core
