+++ core/src/main/scala/com/raphtory/algorithms/generic/TaintTrackingWithHistory.scala
@@ -1,13 +1,10 @@
-package com.raphtory.ethereum.analysis
+package com.raphtory.algorithms.generic
 
-import com.raphtory.api.analysis.algorithm.Generic
 import com.raphtory.api.analysis.algorithm.GenericReduction
-import com.raphtory.api.analysis.graphview.GraphPerspective
-import com.raphtory.api.analysis.graphview.ReducedGraphPerspective
-import com.raphtory.api.analysis.table.Row
-import com.raphtory.api.analysis.table.Table
+import com.raphtory.api.analysis.graphview.{GraphPerspective, ReducedGraphPerspective}
+import com.raphtory.api.analysis.table.{Row, Table}
 
-class Taint(startTime: Long, infectedNodes: Set[String], stopNodes: Set[String] = Set()) extends GenericReduction {
+class TaintTrackingWithHistory(startTime: Long, infectedNodes: Set[String], stopNodes: Set[String] = Set()) extends GenericReduction {
 
   override def apply(graph: GraphPerspective): graph.ReducedGraph =
     graph.reducedView
@@ -157,8 +154,8 @@ class Taint(startTime: Long, infectedNodes: Set[String], stopNodes: Set[String] 
       )
 }
 
-object Taint {
+object TaintTrackingWithHistory {
 
   def apply(startTime: Int, infectedNodes: Set[String], stopNodes: Set[String]) =
-    new Taint(startTime, infectedNodes, stopNodes)
+    new TaintTrackingWithHistory(startTime, infectedNodes, stopNodes)
 }
