--- examples/ethereum/.gitattributes
@@ -1,1 +1,0 @@
-data/transactions_09000000_09099999_small.csv.gz filter=lfs diff=lfs merge=lfs -text
