--- examples/ethereum/lib/readme.md
@@ -1,1 +1,0 @@
-add raphory.jar here
