--- examples/ethereum/project/build.properties
@@ -1,1 +1,0 @@
-sbt.version=1.5.7
