--- examples/ethereum/python/taint_graph.dot
@@ -1,21 +1,0 @@
-digraph "" {
-	startPoint -> "0xa09871aeadf4994ca12f5c0b6056bbd1d343c029";
-	"0xa09871aeadf4994ca12f5c0b6056bbd1d343c029" -> "0x5e032243d507c743b061ef021e2ec7fcc6d3ab89";
-	"0xa09871aeadf4994ca12f5c0b6056bbd1d343c029" -> "0x39925dd94671970b23220cbd5b1e8ed5dd521f15";
-	"0xa09871aeadf4994ca12f5c0b6056bbd1d343c029" -> "0xa09a593804ef2eadce903412983b7f20938597d5";
-	"0xa09871aeadf4994ca12f5c0b6056bbd1d343c029" -> "0x9a207194cbed9f229694fdf5a28caab59157920d";
-	"0xa09871aeadf4994ca12f5c0b6056bbd1d343c029" -> "0x268eac6cf51c5c780b0f68f67c53e9f4e4a06ade";
-	"0xa09871aeadf4994ca12f5c0b6056bbd1d343c029" -> "0xc7d64e6509333a3b68f6fc09d7d19404bfdd229a";
-	"0x5e032243d507c743b061ef021e2ec7fcc6d3ab89" -> "0xa09871aeadf4994ca12f5c0b6056bbd1d343c029";
-	"0x5e032243d507c743b061ef021e2ec7fcc6d3ab89" -> "0x8fd00f170fdf3772c5ebdcd90bf257316c69ba45";
-	"0x5e032243d507c743b061ef021e2ec7fcc6d3ab89" -> "0x39925dd94671970b23220cbd5b1e8ed5dd521f15";
-	"0x39925dd94671970b23220cbd5b1e8ed5dd521f15" -> "0x39925dd94671970b23220cbd5b1e8ed5dd521f15";
-	"0x39925dd94671970b23220cbd5b1e8ed5dd521f15" -> "0xf4678169b1eb0c141fd99942ac02191145fefe3f";
-	"0x39925dd94671970b23220cbd5b1e8ed5dd521f15" -> "0x99ba2bad6da0503e00786375daa9e9554720aed2";
-	"0x39925dd94671970b23220cbd5b1e8ed5dd521f15" -> "0x61a2dee0e8c6fe309f41f2f655d2e72c49b8ae3a";
-	"0x39925dd94671970b23220cbd5b1e8ed5dd521f15" -> "0x3408edca2d47ddaa783a3563d991b8ddebcd973b";
-	"0xa09a593804ef2eadce903412983b7f20938597d5" -> "0xb0ae3f114b6b2727b3ed7e2f6747efc6088db8f9";
-	"0xb0ae3f114b6b2727b3ed7e2f6747efc6088db8f9" -> "0x137ad9c4777e1d36e4b605e745e8f37b2b62e9c5";
-	"0x99ba2bad6da0503e00786375daa9e9554720aed2" -> "0xa450f85dbd67af714ea16a86982223231164fd23";
-	"0xa450f85dbd67af714ea16a86982223231164fd23" -> "0xa11a97912e4ac4d31dce95b2bce0e5cb55f2554a";
-}
