--- examples/ethereum/src/main/resources/ethereum.conf
@@ -1,5 +1,0 @@
-com.raphtory.ethereum {
-    tagFile = "/tmp/etherscan_tags.csv"
-    tagFile = ${?RAPHTORY_ETHEREUM_TAG_FILE}
-}
-
