--- examples/ethereum/src/main/resources/log4j2.xml
@@ -1,49 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Configuration status="INFO" monitorInterval="20">
-    <Appenders>
-        <Console name="Console" target="SYSTEM_OUT">
-            <PatternLayout pattern="%d{HH:mm:ss.SSS} [%t] %-5level %logger{36} - %msg%n" />
-        </Console>
-    </Appenders>
-    <Loggers>
-        <Logger name="org.hibernate.SQL" level="${env:SQL_LOG:-INFO}">
-            <AppenderRef ref="Console"/>
-        </Logger>
-        <Logger name="org.hibernate.type.descriptor.sql" level="${env:SQL_LOG:-INFO}">
-            <AppenderRef ref="Console"/>
-        </Logger>
-        <Logger name="org.hibernate.type.descriptor.sql" level="${env:SQL_LOG:-INFO}">
-            <AppenderRef ref="Console"/>
-        </Logger>
-        <Logger name="com.scurrilous.circe.checksum.Crc32cIntChecksum" level="${env:CHECKSUM_LOG:-off}">
-        </Logger>
-        <Logger name="io.netty" level="${env:NETTY_LOG:-off}">
-        </Logger>
-        <Logger name="org.apache" level="${env:APACHE_LOG:-off}">
-        </Logger>
-        <Logger name="com.raphtory.deployment" level="${env:DEPLOY_LOG:-INFO}">
-        </Logger>
-        <Logger name="com.raphtory.develop" level="${env:DEVELOP_LOG:-INFO}">
-        </Logger>
-        <Logger name="com.raphtory.core" level="${env:CORE_LOG:-INFO}">
-        </Logger>
-        <Logger name="com.raphtory.core.components.querymanager" level="DEBUG">
-        </Logger>
-        <Logger name="com.raphtory.core.components.partition.QueryExecutor" level="DEBUG">
-        </Logger>
-        <Logger name="com.raphtory.core.components.querymanager.QueryManager" level="DEBUG">
-        </Logger>
-
-        <Logger name="com.raphtory.core.components.partition" level="DEBUG">
-        </Logger>
-        <Logger name="com.raphtory.core.components.graphbuilder" level="DEBUG">
-        </Logger>
-        <Logger name="com.raphtory.core.components.spout" level="DEBUG">
-        </Logger>
-
-
-        <Root level="info">
-            <AppenderRef ref="Console"/>
-        </Root>
-    </Loggers>
-</Configuration>
